style: rename files to fix code smell
